Реализовал загрузку списка счетов во время работы сплеш скрина. Реализовал загрузку общего списка транзакций для всех счетов

+++ app/src/main/java/com/example/shmryandex/domain/entity/Expense.kt
@@ -1,9 +1,9 @@
 package com.example.shmryandex.domain.entity
 
 data class Expense(
-    val id: Int,
+    val id: Long,
     val category: Category,
-    val amount: Int,
+    val amount: Double,
     val comment: String = "",
     val createdAt: String
 ) {
